+++ javascript/src/index.ts
@@ -1,19 +1,20 @@
 import Argument from './Argument.js'
 import { Located, Node, NodeType, Token, TokenType } from './Ast.js'
-import CucumberExpression from './CucumberExpression.js'
+import CucumberExpression, { CucumberExpressionJson } from './CucumberExpression.js'
 import CucumberExpressionGenerator from './CucumberExpressionGenerator.js'
 import ExpressionFactory from './ExpressionFactory.js'
 import GeneratedExpression from './GeneratedExpression.js'
 import Group from './Group.js'
-import ParameterType, { RegExps, StringOrRegExp } from './ParameterType.js'
-import ParameterTypeRegistry from './ParameterTypeRegistry.js'
-import RegularExpression from './RegularExpression.js'
+import ParameterType, { ParameterTypeJson, RegExps, StringOrRegExp } from './ParameterType.js'
+import ParameterTypeRegistry, { ParameterTypeRegistryJson } from './ParameterTypeRegistry.js'
+import RegularExpression, { RegularExpressionJson } from './RegularExpression.js'
 import { Expression } from './types.js'
 
 export {
   Argument,
   CucumberExpression,
   CucumberExpressionGenerator,
+  CucumberExpressionJson,
   Expression,
   ExpressionFactory,
   GeneratedExpression,
@@ -22,9 +23,12 @@ export {
   Node,
   NodeType,
   ParameterType,
+  ParameterTypeJson,
   ParameterTypeRegistry,
+  ParameterTypeRegistryJson,
   RegExps,
   RegularExpression,
+  RegularExpressionJson,
   StringOrRegExp,
   Token,
   TokenType,